bug fixes to add post

+++ src/components/Post form/PostForm.jsx
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { useForm } from "react-hook-form";
 
-function PostForm(post) {
+function PostForm({ post }) {
   const { register, handleSubmit, watch, setValue, control, getValues } =
     useForm({
       defaultValues: {
@@ -19,14 +19,12 @@ function PostForm(post) {
   const userData = useSelector((state) => state.auth.userData);
 
   const submit = async (data) => {
-    if (post) {
-      const file = data.images[0]
-        ? await service.uploadFile(data.images[0])
-        : null;
+    if (post && post.$id) {
+      const file = data.image ? await service.uploadFile(data.image) : null;
       if (file) {
         service.deleteFile(post.featuredImage);
       }
-      const dbPost = service.updatePost(ID, {
+      const dbPost = service.updatePost(post.$id, {
         ...data,
         featuredImage: file ? file.$id : undefined,
       });
@@ -34,11 +32,16 @@ function PostForm(post) {
         navigate(`/post/${dbPost.$id}`);
       }
     } else {
-      const file = await service.uploadFile(data.images[0]);
+      console.table(data);
+      const file = await service.uploadFile(data.image[0]);
       if (file) {
         const fileId = file.$id;
         data.featuredImage = fileId;
-        const dbPost = service.createPost({ ...data, userId: userData.$id });
+        const dbPost = service.createPost({
+          ...data,
+          featuredImage: fileId,
+          userId: userData.$id,
+        });
         if (dbPost) {
           navigate(`/post/${dbPost.$id}`);
         }
@@ -47,7 +50,7 @@ function PostForm(post) {
   };
 
   const slugTransform = useCallback((value) => {
-    if (value && typeof value === String) {
+    if (value && typeof value === "string") {
       return value
         .trim()
         .toLowerCase()
@@ -58,7 +61,7 @@ function PostForm(post) {
 
   useEffect(() => {
     const subscription = watch((value, { name }) => {
-      if ((name = "title")) {
+      if (name === "title") {
         setValue("slug", slugTransform(value.title), { shouldValidate: true });
       }
     });
@@ -73,17 +76,17 @@ function PostForm(post) {
           placeholder="Title"
           className="mb-4"
           {...register("title", { required: true })}
+          onInput={(e) => {
+            setValue("slug", slugTransform(e.currentTarget.value), {
+              shouldValidate: true,
+            });
+          }}
         />
         <Input
           label="Slug :"
           placeholder="Slug"
           className="mb-4"
           {...register("slug", { required: true })}
-          onInput={(e) => {
-            setValue("slug", slugTransform(e.currentTarget.value), {
-              shouldValidate: true,
-            });
-          }}
         />
         <RTE
           label="Content :"
